guard with PADDLE_WITH_MKLDNN, tests moved to test_matmul_mkldnn_op.py test=develop
